Trying to create multiple mice
Im trying to figure out a way to add multiple mice right know its just a coppy of the mouseagent brain but I have some ideas how to let a big mouse teach the smaller mice. I dont know when we have to be finished but I think this should be doable before then.